cozinha refactor

+++ src/main/java/dev/ironia/ironeat/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -2,6 +2,7 @@ package dev.ironia.ironeat.infrastructure.repository;
 
 import dev.ironia.ironeat.domain.model.Cozinha;
 import dev.ironia.ironeat.domain.repository.CozinhaRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,8 +35,12 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Override
     @Transactional
-    public void remover(Cozinha cozinha) {
-        cozinha = porId(cozinha.getId());
+    public void remover(Long id) {
+        Cozinha cozinha = porId(id);
+
+        if(cozinha == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cozinha);
     }
 }
